+++ index.html
@@ -14,25 +14,10 @@
 
   </head>
 
-  <body onload="createScene()">
+  <body>
 
-    <div class="container">
-    <canvas id="canvas" height="500" width="800"></canvas>
-
-    <h3>Parameters</h3>
-    <div class="input-group">
-    <span class="input-group-addon">Rim Velocity</span>
-    <input class="form-control" type="range" id="rimvelocity" width="500" min="0" max="0.1"
-	   step="0.001" value="0.02" oninput="onVelocityChanged(this);" ></input>
-   </div>
-
-   <div class="input-group">
-   <span class="input-group-addon">Length</span>
-   <input class="form-control" type="range" id="rimvelocity" width="500" min="10" max="200"
-    step="20" value="100.0" oninput="onLengthChanged(this);" ></input>
-  </div>
-
-  </body>
+    <div id="container"></div>
+
