+++ src/app/components/Pendulum.tsx
@@ -4,6 +4,10 @@ import * as React from "react"
 
 export class Pendulum extends React.Component<undefined, undefined> {
 
-  
+  render() {
+      return (
+          <p>Pendulum</p>
+      )
+  }
 
 }
